requirements.txt
django==4.2.7
djangorestframework>=3.14
django-cors-headers>=4.3
Pillow>=10.4
python-decouple>=3.8
psycopg[binary]>=3.2
